student-review-peer-portal/student-review/requirements.txt
Flask
pymongo
flask-bcrypt
python-dotenv 